# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Imparables SA - Sistema de gestión naturista
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ventas:
    get:
      operationId: listVentas
      tags: [ventas]
      summary: Listar ventas
      responses:
        "200":
          description: Lista de ventas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VentaConDetalle"
    post:
      operationId: createVenta
      tags: [ventas]
      summary: Registrar venta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVentaBody"
      responses:
        "201":
          description: Venta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentaConDetalle"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ventas/{id}:
    get:
      operationId: getVenta
      tags: [ventas]
      summary: Obtener venta por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venta encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentaConDetalle"
        "404":
          description: No encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateVenta
      tags: [ventas]
      summary: Modificar venta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVentaBody"
      responses:
        "200":
          description: Venta actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentaConDetalle"
        "404":
          description: No encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVenta
      tags: [ventas]
      summary: Eliminar venta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venta eliminada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: No encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # REPORTES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    DetalleVenta:
      type: object
      properties:
        idDetalle:
          type: integer
        idProducto:
          type: integer
        nombreProducto:
          type: string
        cantidad:
          type: integer
        subtotal:
          type: number
      required:
        - idDetalle
        - idProducto
        - nombreProducto
        - cantidad
        - subtotal
    VentaConDetalle:
      type: object
      properties:
        idVenta:
          type: integer
        fecha:
          type: string
        idCliente:
          type: integer
        nombreCliente:
          type: string
        idVendedor:
          type: integer
        nombreVendedor:
          type: string
        total:
          type: number
        detalles:
          type: array
          items:
            $ref: "#/components/schemas/DetalleVenta"
      required:
        - idVenta
        - fecha
        - idCliente
        - nombreCliente
        - idVendedor
        - nombreVendedor
        - total
        - detalles
    CreateDetalleVentaItem:
      type: object
      properties:
        idProducto:
          type: integer
        cantidad:
          type: integer
      required:
        - idProducto
        - cantidad
    CreateVentaBody:
      type: object
      properties:
        idCliente:
          type: integer
        idVendedor:
          type: integer
        detalles:
          type: array
          items:
            $ref: "#/components/schemas/CreateDetalleVentaItem"
      required:
        - idCliente
        - idVendedor
        - detalles
...